requirements.txt
python-dotenv
google-generativeai
Pillow
pandas
pdf2image
PyPDF2